update salam files, add lang into ar example files and test files, update dockerfile, bash

+++ compiler/src/interp/interp.c
@@ -37,13 +37,13 @@ void tick(interp_t *I)
 
 static size_t itab_hash(const char *s)
 {
-    size_t h = 1469598103934665603u;
+    uint64_t h = 1469598103934665603ULL;
     const unsigned char *p = (const unsigned char *)s;
     for (; *p; p++) {
         h ^= *p;
-        h *= 1099511628211u;
+        h *= 1099511628211ULL;
     }
-    return h;
+    return (size_t)h;
 }
 
 static size_t ptab_hash(const void *p)
